# [perf] dsv3 bmm fallback to bf16