<!-- Thank you for your contribution! We appreciate it. The following guidelines will help improve your pull request and facilitate feedback. If anything is unclear, don't hesitate to submit your pull request and ask the maintainers for assistance. -->

## Motivation

Align to official impl according to released profile data: https://github.com/deepseek-ai/profile-data/blob/ad8f61ed5e590c4984bb6579f370f9689c427416/decode.json

Issue1(fixed): easyly crash with OOM on H200

fix by #5893 

Issue2(fixed): seems degrade acc on V3-0324 gsm8k. but not affect R1

10 times test after merging with latest main:
```
Accuracy: 0.952
Accuracy: 0.951
Accuracy: 0.955
Accuracy: 0.948
Accuracy: 0.948
Accuracy: 0.945
Accuracy: 0.951
Accuracy: 0.948
Accuracy: 0.948
Accuracy: 0.953
```

remove per group kernel quant in bmm

- deepgemm bs32(2.2 us + 6.9us)
![image](https://github.com/user-attachments/assets/71f4c252-5045-4bd4-8ead-0a8827f36cc1)

- cutlass(5.2us)
![image](https://github.com/user-attachments/assets/31615b44-6ba3-48d4-9dd2-9330cc3c3848)


### MODEL: DeepSeek-V3-0324

#### Before

- 1 batch
```
Prefill. latency: 0.17499 s, throughput:   5851.62 token/s
Decode 0. Batch size: 1, latency: 0.01358 s, throughput:     73.62 token/s
Decode 1. Batch size: 1, latency: 0.01350 s, throughput:     74.07 token/s
Decode 2. Batch size: 1, latency: 0.01346 s, throughput:     74.28 token/s
Decode 3. Batch size: 1, latency: 0.01368 s, throughput:     73.10 token/s
Decode 4. Batch size: 1, latency: 0.01344 s, throughput:     74.43 token/s
Decode.  median latency: 0.01346 s, median throughput:     74.30 token/s
Total. latency:  0.377 s, throughput:   2759.43 token/s
```

- 32 batch
```
Benchmark ...
Prefill. latency: 1.16755 s, throughput:  28065.52 token/s
Decode 0. Batch size: 32, latency: 0.02375 s, throughput:   1347.61 token/s
Decode 1. Batch size: 32, latency: 0.02155 s, throughput:   1485.02 token/s
Decode 2. Batch size: 32, latency: 0.02095 s, throughput:   1527.21 token/s
Decode 3. Batch size: 32, latency: 0.02135 s, throughput:   1499.04 token/s
Decode 4. Batch size: 32, latency: 0.02128 s, throughput:   1503.89 token/s
Decode.  median latency: 0.02253 s, median throughput:   1420.25 token/s
Total. latency:  1.503 s, throughput:  22142.71 token/s
```

#### After

- 1 batch（7.6% boost)
```
Benchmark ...
Prefill. latency: 0.17505 s, throughput:   5849.82 token/s
Decode 0. Batch size: 1, latency: 0.01286 s, throughput:     77.76 token/s
Decode 1. Batch size: 1, latency: 0.01269 s, throughput:     78.77 token/s
Decode 2. Batch size: 1, latency: 0.01257 s, throughput:     79.54 token/s
Decode 3. Batch size: 1, latency: 0.01251 s, throughput:     79.96 token/s
Decode 4. Batch size: 1, latency: 0.01250 s, throughput:     80.00 token/s
Decode.  median latency: 0.01251 s, median throughput:     79.96 token/s
Total. latency:  0.363 s, throughput:   2863.15 token/s
```

- 32 batch (5.5% boost)
```
Benchmark ...
Prefill. latency: 1.17263 s, throughput:  27943.96 token/s
Decode 0. Batch size: 32, latency: 0.02253 s, throughput:   1420.61 token/s
Decode 1. Batch size: 32, latency: 0.02074 s, throughput:   1543.12 token/s
Decode 2. Batch size: 32, latency: 0.02060 s, throughput:   1553.27 token/s
Decode 3. Batch size: 32, latency: 0.02042 s, throughput:   1567.14 token/s
Decode 4. Batch size: 32, latency: 0.02041 s, throughput:   1568.00 token/s
Decode.  median latency: 0.02134 s, median throughput:   1499.39 token/s
Total. latency:  1.495 s, throughput:  22268.03 token/s
```

<!-- Explain the purpose of this PR and the goals it aims to achieve. -->

## Modifications

<!-- Describe the changes made in this PR. -->

## Checklist

- [ ] Format your code according to the [Code Formatting with Pre-Commit](https://docs.sglang.ai/references/contribution_guide.html#code-formatting-with-pre-commit).
- [ ] Add unit tests as outlined in the [Running Unit Tests](https://docs.sglang.ai/references/contribution_guide.html#running-unit-tests-adding-to-ci).
- [ ] Update documentation / docstrings / example tutorials as needed, according to [Writing Documentation](https://docs.sglang.ai/references/contribution_guide.html#writing-documentation-running-docs-ci).
- [ ] Provide throughput / latency benchmark results and accuracy evaluation results as needed, according to [Benchmark and Profiling](https://docs.sglang.ai/references/benchmark_and_profiling.html) and [Accuracy Results](https://docs.sglang.ai/references/accuracy_evaluation.html).
- [ ] For reviewers: If you haven't made any contributions to this PR and are only assisting with merging the main branch, please remove yourself as a co-author when merging the PR.
- [ ] Please feel free to join our Slack channel at https://slack.sglang.ai to discuss your PR.